dom/os/test/test_readdir.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>os readdir test</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1141021">os tests</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>

  <script>
  var jsFuns = SpecialPowers.Cu.getJSTestingFunctions();

  function runTest() {
    return new Promise((res, rej) => {
      var workerCode = `
        try {
          var path = navigator.os.TEMP_DIR + '/test_readdir';
          var files = [path + '/ab', path + '/cd'];
          try {
            navigator.os.unlink(files[0]);
          }
          catch (ex) {}

          try {
            navigator.os.unlink(files[1]);
          }
          catch (ex) {}

          try {
            navigator.os.rmdir(path);
          }
          catch (ex) {}

          function writeFile(path, data) {
            var file = navigator.os.open(path, navigator.os.RDWR | navigator.os.CREAT, 0666);
            var arr = new Uint8Array(data.split('').map(c => c.charCodeAt(0)));
            navigator.os.write(file, arr);
            navigator.os.close(file);
          }

          navigator.os.mkdir(path, navigator.os.IRWXU);
          writeFile(files[0], 'Test1');
          writeFile(files[1], 'Test2');

          var paths1 = navigator.os.readdir(path);
          try {
            navigator.os.readdir(navigator.os.TEMP_DIR + '/nononono');
          }
          catch (ex) {
            var paths2err = ex;
          }

          postMessage(JSON.stringify({
            paths1: paths1,
            paths2err: paths2err
          }));
        }
        catch (ex) {
          postMessage("readdir err: " + ex);
        }
        `;

      var workerBlob = new Blob([workerCode], { type: 'application/javascript' });
      var w = new Worker(URL.createObjectURL(workerBlob));

      w.onmessage = e => {
        is(e.data.indexOf('readdir err'), -1, 'ran successfully');
        if (e.data.indexOf('readdir err') !== -1) {
          dump(e.data + '\n');
          return res();
        }

        try {
          var data = JSON.parse(e.data);
        }
        catch(ex) {
          ok(false, 'is valid JSON (' + e.data + ')');
          return res();
        }

        is(data.paths1.join(','), 'ab,cd', 'paths1');
        is(data.paths2err, 'No such file or directory', 'paths2err');

        res();
      };
    });
  }

  SimpleTest.waitForExplicitFinish();

  Promise.all([
    runTest()
  ]).then(() => {
    SimpleTest.finish();
  });
  </script>
</body>
</html>
